stabilize store merging
Don't let pointer randomization change the order in which we process
store chains.  This may cause SSA_NAMEs to be released in different
order, and if they're reused later, they may cause differences in SSA
partitioning, leading to differences in expand, and ultimately to
different code.

bootstrap-debug-lean (-fcompare-debug) on i686-linux-gnu has failed in
haifa-sched.c since r245196 exposed the latent ordering problem in
store merging.  In this case, the IR differences (different SSA names
selected for copies in out-of-SSA, resulting in some off-by-one
differences in pseudos) were not significant enough to be visible in
the compiler output.


for  gcc/ChangeLog

	* gimple-ssa-store-merging.c (INCLUDE_MAP): Define.
	(struct imm_store_chain_info): Add seqno field and ctor parm.
	(class pass_store_merging): Add m_store_seq field.
	(pass_store_merging::terminate_and_process_all_chains):
	Iterate over m_store_seq.
	(pass_store_merging::terminate_all_aliasing_chains):
	Likewise.
	(pass_store_merging::terminate_and_release_chain):
	Erase the m_store_seq entry.
	(pass_store_merging::execute): Check for debug stmts first.
	Compute seqno, add the m_store_seq entry.